+++ workspaces/orchestrator/plugins/orchestrator-form-api/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @red-hat-developer-hub/backstage-plugin-orchestrator-form-api
 
+## 2.8.3
+
+### Patch Changes
+
+- Updated dependencies [eade824]
+  - @red-hat-developer-hub/backstage-plugin-orchestrator-common@3.7.3
+
+## 2.8.2
+
+### Patch Changes
+
+- Updated dependencies [b2307f3]
+- Updated dependencies [e1a86f0]
+  - @red-hat-developer-hub/backstage-plugin-orchestrator-common@3.7.2
+
 ## 2.8.1
 
 ### Patch Changes